Feature/migrate default values

+++ blog/database/migrations/2019_10_28_132446_create_player_reports_table.php
@@ -14,11 +14,11 @@ class CreatePlayerReportsTable extends Migration
     public function up()
     {
         Schema::create('player_reports', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->timestamps();
             $table->integer('colocacao');
-            $table->integer('player_id');
-            $table->integer('report_id');
+            $table->unsignedBigInteger('player_id');
+            $table->unsignedBigInteger('report_id');
 
             $table->foreign('player_id')
             ->references('id')
